+++ src/main/java/open/commons/ssh/file/IFileDownload.java
@@ -90,7 +90,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      *
      * @return
      *
@@ -115,7 +115,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      * @param overwrite
      *            데이터 저장 경로에 이미 파일이 존재하는 경우 삭제하고 덮어쓸지 여부.<br>
      *            <code>false</code>인 경우 이미 파일이 존재할 때는 {@link FileAlreadyExistsException}을 발생시킨다.
@@ -165,7 +165,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      *
      * @return
      *
@@ -238,7 +238,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      *
      * @return
      *
@@ -263,7 +263,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      * @param overwrite
      *            데이터 저장 경로에 이미 파일이 존재하는 경우 삭제하고 덮어쓸지 여부.<br>
      *            <code>false</code>인 경우 이미 파일이 존재할 때는 {@link FileAlreadyExistsException}을 발생시킨다.
@@ -339,7 +339,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      *
      * @return
      *
@@ -364,7 +364,7 @@ public interface IFileDownload {
      * @param destination
      *            데이터 저장 경로 (절대경로)
      * @param connectTimeout
-     *            접속대기 제한시간 (단위, ms)
+     *            접속대기 제한시간 (단위: ms)
      * @param overwrite
      *            데이터 저장 경로에 이미 파일이 존재하는 경우 삭제하고 덮어쓸지 여부.<br>
      *            <code>false</code>인 경우 이미 파일이 존재할 때는 {@link FileAlreadyExistsException}을 발생시킨다.